backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
python-nmap
dnspython
python-jose